refactor: move base64 related code into its own module
