+++ src/base64.rs
@@ -1,0 +1,5 @@
+mod decoder;
+mod reader;
+
+pub use self::decoder::Base64Decoder;
+pub use self::reader::Base64Reader;
