+++ src/lib.rs
@@ -24,8 +24,7 @@ pub(crate) mod util;
 #[macro_use]
 pub mod errors;
 pub mod armor;
-pub mod base64_decoder;
-pub mod base64_reader;
+pub mod base64;
 pub mod composed;
 pub mod crypto;
 pub mod line_writer;
